OMAP3: PM debug: take PM register snapshot on suspend and resume
On each suspend, save PM register set for ease of PM debugging.

A snapshot of the PRCM registers is taken just before suspend and
available via

  # cat /debug/pm_debug/registers/1

This can be very useful in determing why a given power state was not
hit.

Upon resume, another register snapshot is taken and is available via:

  # cat /debug/pm_debug/registers/2

This can be useful for low-level debugging of wakeup events.

Signed-off-by: Kevin Hilman <khilman@deeprootsystems.com>