temp: OMAP3: PM: GPIO: disable PER UART after GPIO save
FIXME: GPIO core needs to use clock API to prevent this.

When PER UARTs are disabled in idle path and no GPIOs are in use,
the PER block may go idle.  This causes the GPIO context save
that happens right after UART disabling to fail with data aborts
when accessing the GPIO regs in PER.