OMAP3EVM: Fix build errors seen when CONFIG_PM is disabled
Build fails when CONFIG_PM is disabled. This is due to the fact that OPP
tables are defined only when PM is enabled, but is being used in board
file irrespective of whether PM is enabled or disabled.

This is fixed. If PM is not enabled, the board file now marks these
tables as NULL and passes that to omap2_init_common_hw().

Signed-off-by: Ranjith Lohithakshan <ranjithl@ti.com>