OMAP: OPP: Fixing wrong target level being passed during Core DVFS.
As per the current implementaion (u8*)&target_level is being passed to
freq_to_opp in set_opp. This would result in updating just the first 8
bits of a u32 variable. Later target_level is passed to
resource_set_opp_level() as a u32 parameter. This means:

a. Initially target_level was 0xabcdefgh.
b. freq_to_opp updates the lower eight bits of target_level
   to 0xXX. Now target_level = 0xabcdefXX.
c. We pass 0xabcdefXX as target_level to resource_set_opp_level
   when we want to pass just 0xXX.

This is leading to some corrupted bookkeeping later on in the
dvfs path.

This patch ensures that target_level passed to resource_set_opp_level
is actually the level that is intended by freq_to_opp API.

Signed-off-by: Thara Gopinath <thara@ti.com>
Cc: Kevin Hilman <khilman@deeprootsystems.com>