OMAP35xx: Put IVA to idle before rate switching at boot
As a part of the mpu rate switch at boot (using mpurate), the IVA and
core rates are switched as well. This patch puts IVA module to idle
before switching IVA rate. Also, in the case of lowering rates, the
core rate is lowered only after lowering mpu and IVA rates.

This fixes the issue of IVA/CORE/PER powerdomains not entering RET/OFF
states when mpurate is set to 125 (SDOCM00067652).

Signed-off-by: Ranjith Lohithakshan <ranjithl@ti.com>