omap3: Workaround to support 1GHz processor speed
With mpurate=1000 (for the AM/DM37x parts), the kernel
doesn't boot. The operating voltage is not sufficient
for running MPU at 1GHz. Voltage needs to be increased
before increasing the frequency.

I2C subsystem is not initialized when omap2_clk_arch_init()
is called.

This patch implements workaround to delay the frequency
change until SmartReflex driver is initialized to set the
desired frequency.