omap3evm: defconfig: Enable dummy regulator support
Recent changes to ads7846 driver have forced platforms to define a
regulator supply by name "vcc" as the power supply for touchscreen. If
a platform does not implement this, then CONFIG_REGULATOR_DUMMY needs to
be enabled so that a dummy regulator is supplied during regulator api
accesses. If this is not done, the probe and subsequently the suspend
and resume operations can fail.

This patch enables CONFIG_REGULATOR_DUMMY in omap3 evm defconfig.

Signed-off-by: Ranjith Lohithakshan <ranjithl@ti.com>